pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentguard"
version = "0.1.0"
description = "Runtime detection and response platform for AI agents"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "AgentGuard" }]
dependencies = [
    "anthropic>=0.40.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "structlog>=24.4.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-exporter-otlp>=1.28.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.2",
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
openai = ["openai>=1.55.0"]
langgraph = ["langgraph>=0.2.50", "langchain-core>=0.3.25"]
all = ["agentguard[openai,langgraph]"]

[tool.hatch.build.targets.wheel]
packages = ["agentguard"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["agentguard", "api"]
omit = ["tests/*", "examples/*"]

[tool.ruff]
line-length = 100
target-version = "py312"
